Bug in VAT calculation.

+++ basic_webshop/templatetags/vat_tags.py
@@ -7,24 +7,24 @@ register = template.Library()
 
 VAT_PERCENTAGE = getattr(settings, 'WEBSHOP_VAT_PERCENTAGE', 19)
 
-vat_factor = Decimal(VAT_PERCENTAGE/100.0)
+vat_factor = VAT_PERCENTAGE/100.0
 
 
 @register.filter
 def vat_amount(value):
     """ Total amount of VAT for amount. """
-    amount = Decimal(value)
+    amount = float(value)
     return amount*vat_factor
 
 @register.filter
 def vat_inclusive(value):
     """ Amount inlcuding VAT. """
-    amount = Decimal(value)
+    amount = float(value)
     return value*(1+vat_factor)
 
 @register.filter
 def vat_exclusive(value):
     """ Amount inlcuding VAT. """
-    amount = Decimal(value)
+    amount = float(value)
     return value*(1-vat_factor)
